Creating potential field demo.
Potential field provides robot navigation with collision avoidance in a
single interface. This will be used in our strategy module.

This demo, implemented in Python, contains a single robot in a
continuous world where multiple repulsion fields exist. The robot has a
goal position, containing an attraction field, which it must reach
before another goal is assigned. A GUI in Tkinter (using python-tk
package, default in Ubuntu) allows system visualization.

This script is a prototype on how to drive a robotic system using
potential field. Based on this prototype, we can build the system in our
code.